Day 008 - Setup Form and Log View with MongoDB Collection
I made a ton of progress today, I was able to get MongoDB stitch up and running with my app. I really like how easy it was to add a database, I created a form to enter your daily status and a log page to see all of the statuses. Tomorrow, I need to figure out how to clean up some of the collection and only show the latest statuses. Lots to do, but excited!!